requirement.txt
Django==3.0
djangorestframework==3.12.1
fsspec==0.8.3
pandas==1.1.3
pytest==6.1.1
requests==2.24.0
xlrd==1.2.0
